make modal section re-useable and move to its own file
